# Use HTTP Client + CE SDK v2 Bindings in Deliver Processor