--- a/cmd/broker/retry/wire_gen.go
+++ b/cmd/broker/retry/wire_gen.go
@@ -7,7 +7,6 @@ package main
 
 import (
 	"context"
-	"github.com/cloudevents/sdk-go/v2/protocol/http"
 	"github.com/google/knative-gcp/pkg/broker/config/volume"
 	"github.com/google/knative-gcp/pkg/broker/handler/pool"
 )
@@ -23,17 +22,8 @@ func InitializeSyncPool(ctx context.Context, projectID pool.ProjectID, targetsVo
 	if err != nil {
 		return nil, err
 	}
-	v := _wireValue
-	protocol, err := http.New(v...)
-	if err != nil {
-		return nil, err
-	}
-	v2 := _wireValue2
-	deliverClient, err := pool.NewDeliverClient(protocol, v2...)
-	if err != nil {
-		return nil, err
-	}
-	retryPool, err := pool.NewRetryPool(readonlyTargets, client, deliverClient, opts...)
+	httpClient := _wireClientValue
+	retryPool, err := pool.NewRetryPool(readonlyTargets, client, httpClient, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -41,6 +31,5 @@ func InitializeSyncPool(ctx context.Context, projectID pool.ProjectID, targetsVo
 }
 
 var (
-	_wireValue  = pool.DefaultHTTPOpts
-	_wireValue2 = pool.DefaultCEClientOpts
+	_wireClientValue = pool.DefaultHTTPClient
 )