--- a/pkg/broker/handler/pool/fanout.go
+++ b/pkg/broker/handler/pool/fanout.go
@@ -19,6 +19,7 @@ package pool
 import (
 	"context"
 	"fmt"
+	"net/http"
 	"sync"
 	"time"
 
@@ -52,7 +53,7 @@ type FanoutPool struct {
 	deliverRetryClient ceclient.Client
 	// For initial events delivery. We only need a shared client.
 	// And we can set target address dynamically.
-	deliverClient ceclient.Client
+	deliverClient *http.Client
 }
 
 type fanoutHandlerCache struct {
@@ -82,7 +83,7 @@ func (hc *fanoutHandlerCache) shouldRenew(b *config.Broker) bool {
 func NewFanoutPool(
 	targets config.ReadonlyTargets,
 	pubsubClient *pubsub.Client,
-	deliverClient DeliverClient,
+	deliverClient *http.Client,
 	retryClient RetryClient,
 	opts ...Option,
 ) (*FanoutPool, error) {