--- a/pkg/broker/handler/pool/retry.go
+++ b/pkg/broker/handler/pool/retry.go
@@ -19,9 +19,9 @@ package pool
 import (
 	"context"
 	"fmt"
+	"net/http"
 	"sync"
 
-	ceclient "github.com/cloudevents/sdk-go/v2/client"
 	"go.uber.org/zap"
 	"knative.dev/eventing/pkg/logging"
 
@@ -47,7 +47,7 @@ type RetryPool struct {
 	pubsubClient *pubsub.Client
 	// For initial events delivery. We only need a shared client.
 	// And we can set target address dynamically.
-	deliverClient ceclient.Client
+	deliverClient *http.Client
 }
 
 type retryHandlerCache struct {
@@ -74,7 +74,7 @@ func (hc *retryHandlerCache) shouldRenew(t *config.Target) bool {
 }
 
 // NewRetryPool creates a new retry handler pool.
-func NewRetryPool(targets config.ReadonlyTargets, pubsubClient *pubsub.Client, deliverClient DeliverClient, opts ...Option) (*RetryPool, error) {
+func NewRetryPool(targets config.ReadonlyTargets, pubsubClient *pubsub.Client, deliverClient *http.Client, opts ...Option) (*RetryPool, error) {
 	options, err := NewOptions(opts...)
 	if err != nil {
 		return nil, err