Directly use the HTTP client in the deliver processor using the cloudevents SDK v2 bindings. This significantly improves the performance of fanout delivery:

```
name                            old time/op    new time/op    delta
DeliveryNoReply-16                15.0µs ± 1%    11.8µs ± 1%  -21.28%  (p=0.000 n=19+19)
DeliveryWithReply-16              25.8µs ± 1%    23.4µs ± 1%   -9.39%  (p=0.000 n=20+20)
DeliveryNoReplyFakeClient-16      3.03µs ± 1%    1.96µs ± 2%  -35.20%  (p=0.000 n=20+20)
DeliveryWithReplyFakeClient-16    4.88µs ± 2%    3.04µs ± 2%  -37.78%  (p=0.000 n=19+16)

name                            old alloc/op   new alloc/op   delta
DeliveryNoReply-16                8.64kB ± 0%    7.40kB ± 0%  -14.33%  (p=0.000 n=19+18)
DeliveryWithReply-16              17.2kB ± 0%    14.7kB ± 0%  -14.81%  (p=0.000 n=20+20)
DeliveryNoReplyFakeClient-16      4.35kB ± 0%    3.23kB ± 0%  -25.82%  (p=0.000 n=20+20)
DeliveryWithReplyFakeClient-16    7.33kB ± 0%    4.84kB ± 0%  -34.00%  (p=0.000 n=20+16)

name                            old allocs/op  new allocs/op  delta
DeliveryNoReply-16                   116 ± 0%        94 ± 0%  -18.97%  (p=0.000 n=19+20)
DeliveryWithReply-16                 251 ± 0%       210 ± 0%  -16.33%  (p=0.000 n=20+20)
DeliveryNoReplyFakeClient-16        62.0 ± 0%      41.0 ± 0%  -33.87%  (p=0.000 n=20+20)
DeliveryWithReplyFakeClient-16       109 ± 0%        69 ± 0%  -36.70%  (p=0.000 n=20+20)
```