@ian-mi 2020-05-27
I'm not sure exactly why the CE client is so slow. The main performance benefit of the binding code is supposed be removing the intermediate serialization to the event type. I have not yet removed use of the event type from the fanout code, however, and still see significant improvements so I suspect there is some unnecessary overhead in the CE client causing the performance gap.