@n3wscott 2020-05-29
> I wonder what's in CE SDk that is causing the overhead? Is the http protocol not implemented efficiently?

conversion to and from the Event type is a lot of memory copying. Staying with Message and using the Bindings should speed things up for middleware