@ian-mi 2020-06-04
What makes this unexpected is that we haven't yet removed the unnecessary conversion to event and still see a significant performance benefit. I haven't looked too deep into profiling data but I suspect some of this overhead was due to using the client with metrics enabled.